fixed `--languages` option always conflicting with input
As defined in `cli.yml`, `--languages` conflicts with `<input>`,
but if `<input>` has default value, `--languages` will always conflict,
even if no inputs were specified.

Computation of the default value for `<input>` is now moved to `src/main.rs`,
because it cannot be specified in `cli.yml` along with `--languages`.

I suspect this bug was introduced in commit b7a892f4753d, when default
value for `<input>` was added.