add extended LanguageStatisticsOpts, adding file type filtering
This isn't my favorite way of coding this up, but the handlebar code
generation being in a different module makes this a little bit awkward.

This removes building the file walker from utils/fs (and it's
corresponding test, which did not do much) in favor of building the file
walker from a LanguageStatisticsOpts builder.

I've added the opts this way to make it easier to add more options from
the ignore crate later.

This adds a new function, get_statistics_with_opts, to keep backwards
compatibility with the old get_statistics.

The new auto-generated constant creates the full list of supported
languages. In the future, once we get compile time consts, it should be
changed to a hashmap.

The LanguageStatisticsOpts builder is a little bit awkward due to
needing to iterate over all file types for every filtered file type.
A better abstraction (imo) would be to un-public the language modules
and move them around a bit to come up with a cleaner abstraction
overall.

As is, the reality is that processing the files takes far longer than looping
that array, so I'm not too worried.