Enrich the changelog using `cargo changelog`
It looks at the commit history to add statistical information to all
releases, including their dates, while retaining all user edits.

One can and is actually encourated to go in and polish the changelog
before release. If `cargo smart-release` is used for the latter,
it will create tags as with `cargo release`, but additionally
create tag annotations with release notes and create a GitHub release
with release notes as well.

[Gitoxide uses this](https://github.com/Byron/gitoxide/releases) to
maintain changelogs of 20ish crates along with releasing them. The
only manual intervention is to review and potentially polish
the changelogs after they received their scaffolding by parsing
git conventional messages from the git history.

Please feel free to close this PR if this makes no sense here, but
if there is a spark of interest I am happy to explain how doing this
saves time without giving up anything.