Add test of laguage name.
The current generated tests from the data folder check for the
report values but not if the the correct laguage was identified.
The problem here is if the wrong language is identified but by luck
the counts match, the tests wont catch that.

Example:

lets say we have a zsh.zsh file that tokei thinks is a bash script.
Given how simillar they are, the counts will likely be correct even
if the extesion is misenturpreted, but the report produced by tokei
will put the counts under the wrong laguage lable.

A more tricky example is that of ruby_env which relys on the shebang
and the `/usr/bin/env ruby` which is less obvious then some other forms
html